requirements.txt
aiofiles==23.1.0
anyio==3.6.2
autopep8==2.0.1
beanie==1.17.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
lazy-model==0.0.5
motor==3.1.1
pycodestyle==2.10.0
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-multipart==0.0.5
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
